app/requirements.txt
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
grpcio==1.76.0
grpcio-health-checking==1.76.0
grpcio-tools==1.75.1
idna==3.11
protobuf==6.33.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
